Create a geometry group containing multiple geometries

Create a geometry group containing multiple geometries

create_group(name, position = c(0, 0, 0), layer = 1)

Arguments

name
string, name of the geometry

position
x,y,z location of the group

layer
layer of the group. reserved

Value

a GeomGroup instance

Details

A geometry group is a container of multiple geometries. The geometries within the same group share the same shift and rotations (see example 1). In ECoG/iEEG world, you might have 'MRI', 'CT', 'FreeSurfer' that have different orientations. For example, if you want to align MRI to FreeSurfer, Instead of calculating the position of each geometries, you can just put all MRI components into a group, and then set transform of this group, making the group aligned to FreeSurfer.

GeomGroup also can be used to store large data. To generate 3D viewer, `threeBrain` needs to dynamically serialize data into JSON format, which can be read by browsers. However, a FreeSurfer brain might be ~30 MB. This is a very large size and might take ~5 seconds to serialize. To solve this problem, GeomGroup supports cache in its `set_group_data` method. This method supports caching static serialized data into a JSON file, and allows the files to be loaded as static data objects. By "static", I mean the data is not supposed to be dynamic, and it should be "read-only". In JavaScript code, I also optimized such that you don't need to load these large datasets repeatedly. And this allows you to load multiple subjects' brain in a short time.

Examples

# Example 1: relative position

# create group
g = create_group('Group A')

# create two spheres at 10,0,0, but s2 is relative to group A
s1 = geom_sphere('Sphere 1', radius = 2, position = c(10,0,0))
s2 = geom_sphere('Sphere 2', radius = 2, position = c(10,0,0), group = g)

# set transform (rotation)
g$set_transform(matrix(c(
  0,1,0,0,
  1,0,0,0,
  0,0,1,0,
  0,0,0,1
), byrow = TRUE, ncol = 4))

# global position for s2 is 0,10,0
threejs_brain(s1, s2)

# Example 2: cache

# download N27 brain
download_N27_surface(surfaces = 'pial')

#> Error in download_N27_surface(surfaces = "pial"): could not find function "download_N27_surface"

dat = threeBrain:::read_fs_asc('~/rave_data/others/three_brain/N27/std.141.lh.pial.asc')

#> Warning: cannot open file '/Users/beauchamplab/rave_data/others/three_brain/N27/std.141.lh.pial.asc': No such file or directory

#> Error in file(con, "r"): cannot open the connection

vertex = dat$vertices[,1:3]

#> Error in eval(expr, envir, enclos): object 'dat' not found

face = dat$faces[,1:3]

#> Error in eval(expr, envir, enclos): object 'dat' not found

# 1. dynamically serialize
mesh = geom_freemesh('Left Hemisphere', vertex = vertex, face = face)

#> Error in ncol(vertex): object 'vertex' not found

pryr::object_size(mesh) # ~10 MB

#> Registered S3 method overwritten by 'pryr':
#>   method      from
#>   print.bytes Rcpp

#> Error in object_sizes(list(...), env): object 'mesh' not found

threejs_brain(mesh) # ~3 seconds to serialize

#> Error in threejs_brain(mesh): object 'mesh' not found

# 2. cache
tmp_file = tempfile()
mesh = geom_freemesh('Left Hemisphere cached', vertex = vertex,
                     face = face, cache_file = tmp_file)

#> Error in ncol(vertex): object 'vertex' not found

pryr::object_size(mesh) # ~0.5 MB

#> Error in object_sizes(list(...), env): object 'mesh' not found

threejs_brain(mesh) # serialize at once, load in browser

#> Error in threejs_brain(mesh): object 'mesh' not found