Fill in a surface with given volume size

Ray-marching through each margin to find a volume that are inside of a water-tight mesh

fill_surface(
  surface,
  save_as = NULL,
  resolution = 256,
  delta = 2,
  IJK2RAS = NULL,
  verbose = TRUE
)

Arguments

surface

a 'FreeSurfer' surface object, or file path to the object, or a 'rgl' 'mesh3d' object; must be closed, water-tight

save_as

path to file where the results should be saved; the format is 'FreeSurfer' 'mgz'; default is NULL (no save)

resolution

volume resolution; default is to store results in a 256 x 256 x 256 cube

delta

threshold to remove isolated points that are wrongfully created

IJK2RAS

a 4 x 4 volume 'IJK' index to anatomical 'RAS' transform; default is automatically determined

verbose

whether to verbose the progress; default is true

Value

A list of filled volume with IJK2RAS matrix; if save_as is valid, a 'mgz' file will be saved

Examples

library(threeBrain)
if(file.exists(file.path(default_template_directory(), "N27"))) {

  lh_pial <- file.path(default_template_directory(),
                       "N27", "surf", "lh.pial.asc")
  save_as <- tempfile(fileext = ".mgz")

  fill_surface(lh_pial, save_as)

  # you can view `save_as` in FreeSurfer viewer

}