Create a Threejs Brain

Create a Threejs Brain

threejs_brain(..., widget_id = "threebrain_data", time_range = NULL,
  value_range = NULL, symmetric = 0, side_camera = FALSE,
  side_camera_zoom = 1, control_panel = TRUE, control_presets = NULL,
  camera_center = c(0, 0, 0), camera_pos = c(0, 0, 500),
  start_zoom = 1, color_ramp = c("navyblue", "#e2e2e2", "red"),
  color_type = "continuous", n_color = 64,
  color_names = seq_along(color_ramp), show_legend = TRUE,
  legend_title = "Value", tmp_dirname = NULL, width = NULL,
  height = NULL, optionals = list(), debug = FALSE, token = NULL,
  coords = NULL, browser_external = TRUE, global_data = list(),
  .list = list())

Arguments

..., .list
geometries inherit from AbstractGeom

widget_id
unique identifier for the widget. Use it when you have multiple widgets in one website (shiny for example)

time_range
used to calculate animation time (not yet implemented)

value_range
used to generate colors

symmetric
default 0, color center will be mapped to this value

side_camera
enable side cameras to view objects from fixed perspective

side_camera_zoom
if side camera is enabled, zoom-in level, from 1 to 5

control_panel
enable control panels for the widget

control_presets
presets to be shown in control panels

camera_center
position where camera should focus at

camera_pos
XYZ position of camera

start_zoom
numeric, positive number indicating start zoom level

color_ramp
used to generate color ramps

color_type
'continuous' or 'discrete'

n_color
how many colors in the color ramp (used by continuous legend)

color_names
color names (used by discrete legend)

show_legend
show legend in control panel?

legend_title
legend title

tmp_dirname
internally used

width, height
width and height of the widget. By default width="100 and height varies.

optionals
internally used

debug
internally used for debugging

token
used to identify widgets in JS localStorage

coords
NULL to hide coordinates or numeric vector of three.

browser_external
use system default browser (default) or builtin one.

global_data
internally use, mainly to store orientation matrices.