Create a Threejs Brain and View it in Browsers

Create a Threejs Brain and View it in Browsers

threejs_brain(..., .list = list(), width = NULL, height = NULL,
  side_canvas = FALSE, side_zoom = 1, side_width = 250,
  side_shift = c(0, 0), control_panel = TRUE, control_presets = NULL,
  camera_center = c(0, 0, 0), camera_pos = c(0, 0, 500),
  start_zoom = 1, coords = NULL, symmetric = 0,
  default_colormap = "Value", palettes = NULL,
  show_inactive_electrodes = TRUE, widget_id = "threebrain_data",
  tmp_dirname = NULL, debug = FALSE, token = NULL,
  optionals = list(), browser_external = TRUE, global_data = list())

Arguments

..., .list
geometries inherit from AbstractGeom

width, height
positive integers. Width and height of the widget. By default width=`100%`, and height varies.

side_canvas
logical, enable side cameras to view objects from fixed perspective

side_zoom
numerical, if side camera is enabled, zoom-in level, from 1 to 5

side_width
positive integer, side panel size in pixels

side_shift
integer of length two, side panel shift in pixels (`CSS style`: top, left)

control_panel
logical, enable control panels for the widget

control_presets
characters, presets to be shown in control panels

camera_center
numerical, length of three, XYZ position where camera should focus at

camera_pos
XYZ position of camera itself, default (0, 0, 500)

start_zoom
numerical, positive number indicating camera zoom level

coords
NULL to hide coordinates or numeric vector of three.

symmetric
numerical, default 0, color center will be mapped to this value

default_colormap
character, which color map name to display at startup

palettes
named list, names corresponds to color-map names if you want to change color palettes

show_inactive_electrodes
logical, whether to show electrodes with no values

widget_id
character, internally used as unique identifiers for widgets. Only use it when you have multiple widgets in one website

tmp_dirname
character path, internally used, where to store temporary files

debug
logical, internally used for debugging

token
unique character, internally used to identify widgets in JS localStorage

optionals
internally used, to be deprecated

browser_external
logical, use system default browser (default) or builtin one.

global_data
internally use, mainly to store orientation matrices.